requirements.txt
# Root requirements — covers all tools and the setup CLI.
# Install once with: pip install -r requirements.txt

# --- shared across all tools ---
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
anthropic>=0.25.0
pytz
certifi

# --- confluence-migration ---
markdownify>=0.11.6
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-slugify>=8.0.0
rich>=13.0.0

# --- setup CLI ---
click>=8.1.0
questionary>=2.0.0
ruamel.yaml>=0.18.0
